third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.proxy-cross.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify cross-site Partitioned Popins proxies only have access to postMessage and closed methods.
  assert_equals: expected "Closed,Then," but got "Closed,Blur,Opener,Length,Then,"
Harness: the test ran to completion.
